+++ src/chain.js
@@ -1,0 +1,7 @@
+Function.prototype.chain = function() {
+  var me = this;
+  return function() {
+    var ret = me.apply(this, arguments);
+    return ret === undefined ? this : ret;
+  }
+};