+++ test/arboln-test.js
@@ -112,11 +112,33 @@ describe('ArbolN', function(){
 	it('su recorrido postOrden es [befcjklmghida] ', function(){
 	    a.postOrden().should.eql(['b', 'e', 'f', 'c', 'j', 'k', 'l', 'm', 'g', 'h', 'i', 'd', 'a']);
 	});
-	// it('su recorrido por niveles es [abcdefghijklm] ', function(){
-	//     a.niveles().should.eql(['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm']);
-	// });
 	it('el nodo raiz no es hoja', function(){
-	    a.esHoja().should.be.false;
+            a.esHoja().should.be.false;
 	});
+	it('buscamos nodo a y obtenemos a ', function(){
+            a.buscar('a').should.have.property('elemento', 'a');
+	});
+	it('buscamos nodo m y obtenemos m ', function(){
+            a.buscar('m').should.have.property('elemento', 'm');
+	});
+	it('la profundidad de a es 0 ', function(){
+            a.profundidad('a').should.eql(0);
+	});
+	it('la profundidad de m es 3 ', function(){
+            a.profundidad('m').should.eql(3);
+	});
+	it('el padre de a es null', function(){
+	    (a.padreDe('a')===null).should.be.ok;
+	});
+	it('el padre de b es a', function(){
+	    a.padreDe('b').should.have.property('elemento', 'a');
+	});
+	it('el padre de m es g', function(){
+            a.padreDe('m').should.have.property('elemento', 'g');
+	});	
+	
+    // it('su recorrido por niveles es [abcdefghijklm] ', function(){
+    // 	    a.niveles().should.eql(['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm']);
+    // 	});
     });
 });